Pull out node finder into separate file
